Add .gitignore and .h, .cpp files
