+++ include/Device.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+class Device {
+protected:
+    std::string Data;
+
+    std::string generateRandomString(int minLength, int maxLength, const std::vector<char>& validChars);
+
+public:
+    Device();
+
+    virtual void getData() const = 0;
+
+};
