+++ include/DeviceA.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Device.h"
+
+class DeviceA : public Device {
+public:
+    DeviceA();
+
+    void getData() const override;
+};
