+++ include/DeviceB.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Device.h"
+
+class DeviceB : public Device {
+public:
+    DeviceB();
+
+    void getData() const override;
+};
