+++ include/DeviceC.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Device.h"
+
+class DeviceC : public Device {
+public:
+    DeviceC();
+
+    void getData() const override;
+};
